[Decomp][Player] Cohesion split: extract Player::LoadFromDB to PlayerLoadFromDB.cpp
